Backwards incompatible changes to the API
Make a bunch of changes to the API.  The primary goal of the refactoring is to enable composition of HttpClient objects to effect both the RequestBuilder objects and the actual request processing.

There were two functionalities built into HttpClient that fit this mold, one was the credential handling and the other was the "enforce SSL" flag.  I have moved both of them to be wrappers on HttpClient objects.

This is important for me because it will enable me to decorate HttpClient objects with deployment-specific wrappers in Druid that shouldn't muddy up the Druid code proper.